refactor: move createHTTPTransport.
