+++ server/server.go
@@ -38,7 +38,6 @@ import (
 	"github.com/containous/traefik/whitelist"
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/negroni"
-	"golang.org/x/net/http2"
 )
 
 var httpServerLogger = stdlog.New(log.WriterLevel(logrus.DebugLevel), "", 0)
@@ -133,79 +132,6 @@ func NewServer(globalConfiguration configuration.GlobalConfiguration, provider p
 		}
 	}
 	return server
-}
-
-type h2cTransportWrapper struct {
-	*http2.Transport
-}
-
-func (t *h2cTransportWrapper) RoundTrip(req *http.Request) (*http.Response, error) {
-	req.URL.Scheme = "http"
-	return t.Transport.RoundTrip(req)
-}
-
-// createHTTPTransport creates an http.Transport configured with the GlobalConfiguration settings.
-// For the settings that can't be configured in Traefik it uses the default http.Transport settings.
-// An exception to this is the MaxIdleConns setting as we only provide the option MaxIdleConnsPerHost
-// in Traefik at this point in time. Setting this value to the default of 100 could lead to confusing
-// behaviour and backwards compatibility issues.
-func createHTTPTransport(globalConfiguration configuration.GlobalConfiguration) *http.Transport {
-	dialer := &net.Dialer{
-		Timeout:   configuration.DefaultDialTimeout,
-		KeepAlive: 30 * time.Second,
-		DualStack: true,
-	}
-	if globalConfiguration.ForwardingTimeouts != nil {
-		dialer.Timeout = time.Duration(globalConfiguration.ForwardingTimeouts.DialTimeout)
-	}
-
-	transport := &http.Transport{
-		Proxy:                 http.ProxyFromEnvironment,
-		DialContext:           dialer.DialContext,
-		MaxIdleConnsPerHost:   globalConfiguration.MaxIdleConnsPerHost,
-		IdleConnTimeout:       90 * time.Second,
-		TLSHandshakeTimeout:   10 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-	}
-
-	transport.RegisterProtocol("h2c", &h2cTransportWrapper{
-		Transport: &http2.Transport{
-			DialTLS: func(netw, addr string, cfg *tls.Config) (net.Conn, error) {
-				return net.Dial(netw, addr)
-			},
-			AllowHTTP: true,
-		},
-	})
-
-	if globalConfiguration.ForwardingTimeouts != nil {
-		transport.ResponseHeaderTimeout = time.Duration(globalConfiguration.ForwardingTimeouts.ResponseHeaderTimeout)
-	}
-	if globalConfiguration.InsecureSkipVerify {
-		transport.TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
-	}
-	if len(globalConfiguration.RootCAs) > 0 {
-		transport.TLSClientConfig = &tls.Config{
-			RootCAs: createRootCACertPool(globalConfiguration.RootCAs),
-		}
-	}
-	http2.ConfigureTransport(transport)
-
-	return transport
-}
-
-func createRootCACertPool(rootCAs traefiktls.RootCAs) *x509.CertPool {
-	roots := x509.NewCertPool()
-
-	for _, cert := range rootCAs {
-		certContent, err := cert.Read()
-		if err != nil {
-			log.Error("Error while read RootCAs", err)
-			continue
-		}
-		roots.AppendCertsFromPEM(certContent)
-	}
-
-	return roots
 }
 
 // Start starts the server.
